+++ akshare/__init__.py
@@ -2154,9 +2154,10 @@ amac_manager_cancelled_info # 中国证券投资基金业协会-信息公示-诚
 1.7.34 fix: fix bond_cb_jsl interface
 1.7.35 fix: fix stock_market_fund_flow interface
 1.7.36 fix: fix rename branch master to main
+1.7.37 add: add sw_index_first_info interface
 """
 
-__version__ = "1.7.36"
+__version__ = "1.7.37"
 __author__ = "AKFamily"
 
 import sys
@@ -3831,6 +3832,8 @@ from akshare.index.index_sw import (
     sw_index_daily,
     sw_index_daily_indicator,
     sw_index_third_cons,
+    sw_index_first_info,
+    sw_index_second_info,
     sw_index_third_info,
     index_level_one_hist_sw,
     index_style_index_hist_sw,